Whoops, the beneficiary count type was causing test failures. Fixed.
CBBD-10.

+++ bluebutton-data-pipeline-desynpuf/src/main/java/gov/hhs/cms/bluebutton/datapipeline/desynpuf/SynpufArchive.java
@@ -45,7 +45,7 @@ public enum SynpufArchive {
 	 * @return the number of unique beneficiaries stored in this
 	 *         {@link SynpufArchive}
 	 */
-	public int getBeneficiaryCount() {
+	public long getBeneficiaryCount() {
 		return beneficiaryCount;
 	}
 }
